Add divider and timer common modules
Added divider module, which is meant to be used to divide clock
signal by an integer to slowed down enable signal. One can use a
tree of them to generate several slowed enable signals from one
clock. NOTE: the divided signal is not to be used as clock as it
has delay related to original clock and would create a separate
clock domain in effect.

Added timer module, which takes a delay in milliseconds and a 1kHz
slowed clock signal. When enable is set high the delay is loaded
and the timer starts counting down to zero. When 0 a timeout signal
is enabled for one clock cycle and counting is finished.